disease_diagnose/requirements.txt
# Function dependencies, for example:
# package>=version
numpy
pandas
matplotlib
seaborn
sklearn
firebase-admin
google-cloud-storage